extract PlansSetupDialog component
